fix: prop handling on native Android implementation
